requirements.txt
requests
mysql-connector-python
oauth2client
google-cloud-bigquery
pandas-gbq
gspread
google-auth-oauthlib
google-api-python-client
google-auth-httplib2